Implement diffEQ logistic growth model
The `reproduce()` function now implements the equation `P(t) = -Kce^(rt)/(1+ce^(rt))`, where `K`  is the carrying capacity, `c` determines the initial condition, and `r` is a constant determining the speed of growth. This equation is a solution to the differential equation `dP/dt = rP(1-(P/K))`, which is the standard differential equation used to model logistic growth. 

Every season, each lake determines the value of `c` such that `P(0) = self.population`. Next, it plugs in this `c` value along with `self.carrying_capacity` and the constant `r` into the equation, and sets `t` to 1. This yields the new population value.

This should solve issue #3, and is preferable over the previous "iterate a whole bunch of times every season" approach both because it's more computationally efficient and because it yields consistent results. (The results of the model under the previous approach did not appear to converge as `reproductive_resolution` approached infinity.)